requirements.txt
# Core OCR dependencies
pytesseract>=0.3.10
pdf2image>=1.16.3
Pillow>=9.5.0
openpyxl>=3.1.0

# PaddleOCR support (optional but recommended for better Chinese recognition)
paddleocr>=2.7.0
paddlepaddle>=2.5.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.0

# Type checking and code quality
mypy>=1.5.0
black>=23.7.0
flake8>=6.1.0

# Utilities
tqdm>=4.66.0  # Progress bars
colorama>=0.4.6  # Cross-platform colored terminal text
